get detail

+++ app/src/main/java/com/team2/unithon11th_team2_android/features/map/MapNavigation.kt
@@ -10,7 +10,7 @@ const val MAP_ROUTE = "map"
 fun NavHostController.navigateToMapScreen() = navigate(MAP_ROUTE)
 
 @SuppressLint("MissingPermission")
-fun NavGraphBuilder.mapScreen(navigateToRespond: () -> Unit, finish: () -> Unit){
+fun NavGraphBuilder.mapScreen(navigateToRespond: (Int) -> Unit, finish: () -> Unit){
     composable(MAP_ROUTE) {
         MapScreen(navigateToRespond = navigateToRespond, onBackPressed = finish)
     }
